+++ includes/handlers/form/src/FormValuesManager.php
@@ -2,6 +2,7 @@
 
 namespace AKlump\CheckPages\Handlers\Form;
 
+use AKlump\CheckPages\Exceptions\DeprecatedSyntaxException;
 use InvalidArgumentException;
 
 class FormValuesManager {
@@ -10,11 +11,13 @@ class FormValuesManager {
 
   const OPTION_BLOCK_NON_FORM_KEYS = 1;
 
-  private array $config;
+  private array $config = [];
 
-  private array $formValues;
+  private array $formValues = [];
 
   private int $options;
+
+  private array $allowedValues;
 
   /**
    * @param int $options
@@ -29,21 +32,24 @@ class FormValuesManager {
     $this->formValues = $form_values;
   }
 
+  public function setAllowedValues(array $allowed_values): void {
+    $this->allowedValues = $allowed_values;
+  }
+
   public function setConfig(array $config): void {
     if (!empty($config['form']['input'])) {
       if (!is_numeric(key($config['form']['input']))) {
         throw new InvalidArgumentException('form.input must be an indexed array');
       }
-      foreach ($config['form']['input'] as $value) {
-        if (!isset($value['name'])) {
-          throw new InvalidArgumentException('form.input[*].name is missing');
+      foreach ($config['form']['input'] as $index => $value) {
+        if (empty($value['name'])) {
+          throw new InvalidArgumentException("form.input[$index].name is missing");
         }
-        $value_key = array_intersect_key($value, [
-          'value' => 1,
-          'option' => 1,
-        ]);
-        if (empty($value_key)) {
-          throw new InvalidArgumentException('form.input[*].value || form.input[*].option is required');
+        if (array_key_exists('option', $value)) {
+          throw new DeprecatedSyntaxException("form.input[$index].option", "form.input[$index].value");
+        }
+        if (!array_key_exists('value', $value)) {
+          throw new InvalidArgumentException("form.input[$index].value is required");
         }
       }
     }
@@ -71,17 +77,48 @@ class FormValuesManager {
     if (empty($config['form']['input'])) {
       return [];
     }
-    foreach ($config['form']['input'] as $data) {
+    foreach ($config['form']['input'] as $index => $data) {
       $name = $data['name'];
-      $value = $data['value'] ?? $data['option'] ?? NULL;
-      if (isset($this->formValues[$name])
+      $value = $data['value'];
+
+      // Make sure the value is allowed.
+      if (isset($this->allowedValues[$name])) {
+        $is_allowed = FALSE;
+        foreach ($this->allowedValues[$name] as $allowed_value) {
+          $matched_value = $this->matchConfigValueToFormValue($value, $allowed_value);
+          if (NULL !== $matched_value) {
+            $is_allowed = TRUE;
+            $value = $matched_value;
+            break;
+          }
+        }
+        if (!$is_allowed) {
+          throw new InvalidArgumentException("form.input[$index].value is one of allowed values: %s", implode(', ', $this->allowedValues[$name]));
+        }
+      }
+      elseif (isset($this->formValues[$name])
         && $this->formValues[$name] instanceof KeyLabelNode) {
-        $this->formValues[$name]->mutateToKey($value);
+        $matched_value = $this->matchConfigValueToFormValue($value, $this->formValues[$name]);
+        if (NULL === $matched_value) {
+          throw new InvalidArgumentException(sprintf('Config value form.input[%d][%s]=%s is not a valid form key or case-insensitive label matching "%s"', $index, $name, $value, $this->formValues[$name]));
+        }
+        $value = $matched_value;
       }
       $values[$name] = $value;
     }
 
     return $values;
+  }
+
+
+  private function matchConfigValueToFormValue($value, KeyLabelNode $form_input) {
+    $key = $form_input->getKey();
+    if ($value == $key) {
+      return $key;
+    }
+    if (strcasecmp($form_input->getLabel(), $value) === 0) {
+      return $key;
+    }
   }
 
 }